feature: add Cache functionality for Assets
