+++ CybridSDKTests/Mocks/Prices/PricesAPIMock.swift
@@ -28,22 +28,3 @@ final class PricesAPIMock: PricesAPI {
     listPricesCompletion?(.failure(.error(0, nil, nil, CybridError.serviceError)))
   }
 }
-
-extension Array where Element == SymbolPriceBankModel {
-  static let mockPrices: Self = [
-    SymbolPriceBankModel(
-      symbol: "BTC-USD",
-      buyPrice: 2_019_891,
-      sellPrice: 2_019_881,
-      buyPriceLastUpdatedAt: nil,
-      sellPriceLastUpdatedAt: nil
-    ),
-    SymbolPriceBankModel(
-      symbol: "ETH-USD",
-      buyPrice: 209_891,
-      sellPrice: 209_881,
-      buyPriceLastUpdatedAt: nil,
-      sellPriceLastUpdatedAt: nil
-    )
-  ]
-}
